@@ -2,28 +2,27 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace DiaryScheduler.Data.Models
+namespace DiaryScheduler.Data.Models;
+
+public class CalendarEvent
 {
-    public class CalendarEvent
-    {
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public Guid CalendarEventId { get; set; }
+    [Key]
+    [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+    public Guid CalendarEventId { get; set; }
 
-        [Required]
-        public string UserId { get; set; }
-        public DateTime DateFrom { get; set; }
-        public DateTime DateTo { get; set; }
-        public bool AllDay { get; set; }
+    [Required]
+    public string UserId { get; set; }
+    public DateTime DateFrom { get; set; }
+    public DateTime DateTo { get; set; }
+    public bool AllDay { get; set; }
 
-        [Required]
-        [StringLength(100)]
-        public string Title { get; set; }
+    [Required]
+    [StringLength(100)]
+    public string Title { get; set; }
 
-        [StringLength(200)]
-        public string Description { get; set; }
+    [StringLength(200)]
+    public string Description { get; set; }
 
-        [ForeignKey("UserId")]
-        public virtual ApplicationUser AspNetUser { get; set; }
-    }
+    [ForeignKey("UserId")]
+    public virtual ApplicationUser AspNetUser { get; set; }
 }
