@@ -1,38 +1,37 @@
-﻿namespace DiaryScheduler.Presentation.Models.Scheduler
+﻿namespace DiaryScheduler.Presentation.Models.Scheduler;
+
+/// <summary>
+/// The scheduler index view model.
+/// </summary>
+public class SchedulerIndexViewModel
 {
     /// <summary>
-    /// The scheduler index view model.
+    /// Gets or sets the create event url.
     /// </summary>
-    public class SchedulerIndexViewModel
-    {
-        /// <summary>
-        /// Gets or sets the create event url.
-        /// </summary>
-        public string CreateEventUrl { get; set; }
+    public string CreateEventUrl { get; set; }
 
-        /// <summary>
-        /// Gets or sets the create event with more options url.
-        /// </summary>
-        public string CreateEventMoreOptionsUrl { get; set; }
+    /// <summary>
+    /// Gets or sets the create event with more options url.
+    /// </summary>
+    public string CreateEventMoreOptionsUrl { get; set; }
 
-        /// <summary>
-        /// Gets or sets the edit event url.
-        /// </summary>
-        public string EditEventUrl { get; set; }
+    /// <summary>
+    /// Gets or sets the edit event url.
+    /// </summary>
+    public string EditEventUrl { get; set; }
 
-        /// <summary>
-        /// Gets or sets the calendar source url.
-        /// </summary>
-        public string CalendarSourceUrl { get; set; }
+    /// <summary>
+    /// Gets or sets the calendar source url.
+    /// </summary>
+    public string CalendarSourceUrl { get; set; }
 
-        /// <summary>
-        /// Gets or sets the create event url.
-        /// </summary>
-        public string PostCreateEventUrl { get; set; }
+    /// <summary>
+    /// Gets or sets the create event url.
+    /// </summary>
+    public string PostCreateEventUrl { get; set; }
 
-        /// <summary>
-        /// Gets or sets the export ical url.
-        /// </summary>
-        public string ExportIcalUrl { get; set; }
-    }
+    /// <summary>
+    /// Gets or sets the export ical url.
+    /// </summary>
+    public string ExportIcalUrl { get; set; }
 }
