@@ -1,20 +1,19 @@
 ﻿using System;
 
-namespace DiaryScheduler.Presentation.Services.Utility
-{
-    /// <summary>
-    /// The implementation of the <see cref="IDateTimeService"/>.
-    /// </summary>
-    public class DateTimeService : IDateTimeService
-    {
-        public DateTime GetDateTimeNow()
-        {
-            return DateTime.Now;
-        }
+namespace DiaryScheduler.Presentation.Services.Utility;
 
-        public DateTime GetDateTimeUtcNow()
-        {
-            return DateTime.UtcNow;
-        }
+/// <summary>
+/// The implementation of the <see cref="IDateTimeService"/>.
+/// </summary>
+public class DateTimeService : IDateTimeService
+{
+    public DateTime GetDateTimeNow()
+    {
+        return DateTime.Now;
+    }
+
+    public DateTime GetDateTimeUtcNow()
+    {
+        return DateTime.UtcNow;
     }
 }
