@@ -1,45 +1,44 @@
 ﻿using System;
 
-namespace DiaryScheduler.ScheduleManagement.Core.Models
+namespace DiaryScheduler.ScheduleManagement.Core.Models;
+
+/// <summary>
+/// The calendar event domain model.
+/// </summary>
+public class CalEventDm
 {
     /// <summary>
-    /// The calendar event domain model.
+    /// Gets or sets the calendar entry id.
     /// </summary>
-    public class CalEventDm
-    {
-        /// <summary>
-        /// Gets or sets the calendar entry id.
-        /// </summary>
-        public Guid CalendarEntryId { get; set; }
+    public Guid CalendarEntryId { get; set; }
 
-        /// <summary>
-        /// Gets or sets the user id.
-        /// </summary>
-        public string UserId { get; set; }
+    /// <summary>
+    /// Gets or sets the user id.
+    /// </summary>
+    public string UserId { get; set; }
 
-        /// <summary>
-        /// Gets or sets the from date.
-        /// </summary>
-        public DateTime DateFrom { get; set; }
+    /// <summary>
+    /// Gets or sets the from date.
+    /// </summary>
+    public DateTime DateFrom { get; set; }
 
-        /// <summary>
-        /// Gets or sets the to date.
-        /// </summary>
-        public DateTime DateTo { get; set; }
+    /// <summary>
+    /// Gets or sets the to date.
+    /// </summary>
+    public DateTime DateTo { get; set; }
 
-        /// <summary>
-        /// Gets or sets a value indicating whether the entry is an all day event.
-        /// </summary>
-        public bool AllDay { get; set; }
+    /// <summary>
+    /// Gets or sets a value indicating whether the entry is an all day event.
+    /// </summary>
+    public bool AllDay { get; set; }
 
-        /// <summary>
-        /// Gets or sets the calendar event title.
-        /// </summary>
-        public string Title { get; set; }
+    /// <summary>
+    /// Gets or sets the calendar event title.
+    /// </summary>
+    public string Title { get; set; }
 
-        /// <summary>
-        /// Gets or sets the calendar event description.
-        /// </summary>
-        public string Description { get; set; }
-    }
+    /// <summary>
+    /// Gets or sets the calendar event description.
+    /// </summary>
+    public string Description { get; set; }
 }
